Added the option for the user to opt out of saving the CSV file, added the option for the user to input the file path for the CSV file to be saved

+++ qualifier/output/save_to_csv.py
@@ -1,5 +1,8 @@
 import csv
 from pathlib import Path
+import sys
+import questionary
+import os
 
 def save_qualifying_loans(qualifying_loans):
     """Saves the qualifying loans to a CSV file.
@@ -7,13 +10,22 @@ def save_qualifying_loans(qualifying_loans):
     Args:
         qualifying_loans (list of lists): The qualifying bank loans.
     """
-    header = ["Lender", "Max Loan Amount", "Max LTV", "Max DTI", "Min Credit Score", "Interest Rate"]
-    csvpath = Path("qualifying_loans.csv")
-    with open(csvpath, 'w', newline= '') as csvfile:
-        csvwriter = csv.writer(csvfile)
-        csvwriter.writerow(header)
-        for row in qualifying_loans:
-            csvwriter.writerow(row)
-    print("Saving qualifying loans to a CSV file...")
-    return csvfile
     
+    if len(qualifying_loans) > 0:
+        save_loan = questionary.confirm("Would you like to save qualifying loans to a CSV file?").ask()
+        if save_loan == True:
+            desired_path = questionary.text("Enter a file path to the folder where you would like to save your file:").ask()
+            desired_path = Path(desired_path)
+            if not desired_path.exists():
+                sys.exit(f"Oops! Can't find this path: {desired_path}")
+            header = ["Lender", "Max Loan Amount", "Max LTV", "Max DTI", "Min Credit Score", "Interest Rate"]
+            csvpath = os.path.join(desired_path, "qualifying_loans.csv")
+            with open(csvpath, 'w', newline= '') as csvfile:
+                csvwriter = csv.writer(csvfile)
+                csvwriter.writerow(header)
+                for row in qualifying_loans:
+                    csvwriter.writerow(row)
+            print("Saving qualifying loans to a CSV file...")
+            return csvfile
+    else:
+        sys.exit("Sorry, you do not qualify for any loans")
